changed uint8_t to uint32_t for wake up period.

+++ STM32Sleep.h
@@ -6,7 +6,7 @@ typedef enum SleepMode {
   STANDBY
 } SleepMode;
 
-void sleepAndWakeUp(SleepMode mode, RTClock *rt, uint8_t seconds);
+void sleepAndWakeUp(SleepMode mode, RTClock *rt, uint32_t seconds);
 
 void goToSleep(SleepMode mode);
